resources: Update README with Makefile info
This build information was previously in the Makefile (now removed).
This commit adds it to the root README file.

Note: The layout and information in the root README file needs updated.
Some of the information may be better added to subdirectory READMEs.
The purpose of this commit is to add build information so it is
available, at least, somewhere.

Change-Id: I29db36f5016b36484370ddfba59d317b3533c54d
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/32194
Reviewed-by: Hoa Nguyen <hoanguyen@ucdavis.edu>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>