resources: Fix PARSEC compiling issues
libtool wasn't built correctly as aclocal, automake and autoconf
was not installed. As a result, some of the workloads were not
able to be built.

This patch fixes this issue by install autotools-dev and
autoconf.

Also, the patch adds change the owner of the parsec-benchmark
folder to the user 'gem5', which will run the workloads.

JIRA: https://gem5.atlassian.net/browse/GEM5-723

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>
Change-Id: I12e756ac5620ce7e92b7e403bf45f87723ed3164
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/33414
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>