resources: Removed the dependencies section
The number of dependencies to compile/create all the gem5 resources is
growing, and as such, it's better to which are needed for each project.
In this commit the "Dependencies" section is removed, and the
dependencies are stated within each gem5 resource's blurb.

Change-Id: Ib3f68858d347a086e6879a972e046c7e8add7b05
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/40039
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Reviewed-by: Hoa Nguyen <hoanguyen@ucdavis.edu>