resources: print a statement to std when booting Linux is done
This is useful for checking the status of booting Linux with KVM.
Occasionally, booting Linux kernel with KVM gets stuck and there
is no mechanism to detect this. By expecting the simulation
to print out a statement indicating the booting process is done
after a reasonable amount of time, one can detect this issue
automatically.

In other words, this is useful to implement timeout specifically
for booting Linux kernel with KVM.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>
Change-Id: Ieddd28a946857528f73e58636ff1de184ff73702
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/44125
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>