resources: Add resources.json
The purpose of resources.json is to provide a machine-readable lookup
for the gem5 resources, where they may be downloaded, where there
sources exist, and other relevant information.

At present only a small subset of the resources has been added for
testing purposes. This will be expanded when this current schema is
found to be suitable.

Change-Id: If216194ca1c65975fa72b163612d4ce0c5ecba27
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/47700
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>