resources: Update m5 command to cross compile
By default, the cross compiler used by scons to compile riscv targets
is riscv64-unknown-linux-gnu-, which is the one used throughout this
README file. However, it's important to point out how to change
the cross compiler as it's very likely to be varied from a system
to another.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>
Change-Id: Iba29b7e38fbb5c065459c4d229929350bdce90ec
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/48240
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>