resources: Update heterosync for ROCm 4 and dGPU support
This patch implements workarounds for the memory management bugs that
exist in gem5 as a result of the update to ROCm 4, as well as updates to
the Makefile to build for both supported GPU versions in the same
binary.

The patch replaces calls to hipMalloc with hipHostMalloc, as
hipHostMalloc creates host and device memory at the same time without
needing to use hipMemcpy, which currently doesn't work for hipMalloc'd
memory in gem5.

This patch also updates the Makefile to build for both gfx801 and gfx803
in the same binary

Change-Id: I0e289fa89baefa09ba360c151f9c7ba70041d961
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/48484
Reviewed-by: Matt Sinclair <mattdsinclair@gmail.com>
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Matt Sinclair <mattdsinclair@gmail.com>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>