resources: Update DNNMark for ROCm 4 and dGPU support
This patch implements fixes for building DNNMark, building MIOpen
cachefiles, and workarounds for memory management bugs in gem5 that
were a result of the upgrade to ROCm 4.

Previously, DNNMark used the hcc compiler, which is now removed from
ROCm. This patch updates DNNMark's build process to use hipcc instead.

Previously, MIOpen cachefiles were located in a directory that contained
all of the files individually. ROCm 4 instead uses a SQLite database.
This patch removes the old method used for generating cachefiles, and
implements a new method in python that handles creating and inserting
the cachefiles into the database.

This patch also updates the memory management code in DNNMark to use
hipHostMalloc instead of hipMalloc, as hipMemcpys on hipMalloc'd memory
is currently not working in gem5.

This patch also removes the Dockerfile previously used to create an
image with glog and gflags, as they are installed by default in the
current gcn-gpu Docker image.

Change-Id: I3045fbd2adfbf69904732a52f04d963c0cd62482
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/48485
Reviewed-by: Matt Sinclair <mattdsinclair@gmail.com>
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Matt Sinclair <mattdsinclair@gmail.com>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>