resources: Update HACC for ROCm 4.0 and gfx803
This patch updates HACC to build under ROCm 4 and Ubuntu 20 by updating
the appropriate environment variables in the Dockerfile to account for
both the updates to ROCm paths and MPI/OMP paths.

Additionally, this sets the HCC_AMDGPU_TARGET environment variable to
build HACC for both gfx801 and gfx803 at the same time.

Finally, this patch updates the HACC source code to be compatible with
ROCm 4 and the bugs that exist between gem5 and ROCm 4 by replacing
mallocs and hipMallocs with hipHostMalloc, and hipDeviceSynchronize with
hipStreamSynchronize

The output has been validated against and matches a run done on a
real GPU

Change-Id: I91929abd3fdb58ff2a022b3f41c0053435c183ba
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/50987
Reviewed-by: Matt Sinclair <mattdsinclair@gmail.com>
Reviewed-by: Matthew Poremba <matthew.poremba@amd.com>
Reviewed-by: Alex Dutu <alexandru.dutu@amd.com>
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Matt Sinclair <mattdsinclair@gmail.com>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>