resources: Move url_base from the url in resources.json
The gem5 library will now process the URL as an f-string to replace
"url_base" which what's set at the top of the file.

This allows us to update the URLs easier for releases, and in the
eventuality the resources are moved to another url.

Change-Id: Ic7d4e8cc74019d2e20dea51a1f70b4cc1de65b1d
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/51167
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>