resources: Update 'riscv-lupio-busybox-img' description
This patch updates the fixes the incorrect description for the
'riscv-lupio-busybox-img' LupV disk image resources. The patch removed
reference to a non-existence "LupioBoard" (it's the "Lupvboard"), and
fixes the incorrect statement that it must be used with the
'riscv-lupio-busybox-img' resource (it must be used with the
'riscv-lupio-linux-kernel' resource).

Change-Id: Iebb2f04a110a20d835d9c8b6820da6fb6ae2cb3f
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/56570
Reviewed-by: Bobby Bruce <bbruce@ucdavis.edu>
Maintainer: Bobby Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby Bruce <bbruce@ucdavis.edu>