resources: Fix RISCV Tests compilation instructions in README.md
The "RISCV Tests" instructions are missing the compiler prefix for the
make command. The RISCV_PREFIX variable must be passed via the
command line. This variable indicates the prefix of the name of the
RISCV compiler binary on the local machine. The prefix is all
characters that precede "gcc". For example, if the compiler binary's
name is "riscv64-linux-gnu-gcc", then the prefix is
"riscv64-linux-gnu-". This is indicated on the README.md with an
example on how it is used. These commands were tested on the
llvm-gnu-cross-compiler-riscv64 docker container with the prefix
riscv64-unknown-linux-gnu-. The Dockerfile for this image can be
found at util/dockerfiles/llvm-gnu-cross-compiler-riscv64 in the gem5
repository.

Change-Id: I8683c4caf7078c984e4f077ea533d680c1a1572f
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/61450
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>