resources: Update riscv-ubuntu README
This change updates the README to more accurately reflect the
directory structure of src/riscv-ubuntu.

Instructions on inserting the provided exit.sh into the image have
been added.

Made clarification on the U-boot version that must be used in qemu.
A more update to date download link for U-boot is provided.

Added instructions on how to boot the locally built image in gem5.

Change-Id: I1027712924b19edaf2d0eee39cbbb966e3272e17
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/65391
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>
Reviewed-by: Bobby Bruce <bbruce@ucdavis.edu>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>