resources: Update arm-ubuntu README
Corrected the Ubuntu package names for the arm cross compilers.

Added instructions for generating the SSH key pair.

Added instructions on connecting to the QEMU VM with SSH if the
host username is different from what is printed by shell command
"whoami". An example of where this is necessary is when building the
disk image using DArchR computers.

Clarified that cloud-init should finish upon boot in QEMU before
an SSH connection can be made. This usually happens some time after
the VM displays the log in prompt.

Changed the build.sh error messages to specify what caused the error.

The build.sh script currently checks if "name" is set to "ubuntu" to
detect if cloud.txt is unmodified. Added "ubuntu" as the name to the
default provided cloud.txt.

Change-Id: I06c3d2118706886aab3566e4a318a92fd4a72f30
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-resources/+/65471
Tested-by: Jason Lowe-Power <power.jg@gmail.com>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>