resources: Add x86-gapbs on Ubuntu 22.04 disk image resource
This commit adds the files necessary to create an x86 Ubuntu 22.04 disk image containing the GAP
Benchmark Suite. It only includes the files necessary to create the disk image, and not the disk
image itself. The files in the directory "qemu_files" are used to create the disk image and launch
a QEMU instance. The directory "packer_files" contains the files needed by Packer to install GAPBS
onto the disk image. The README contains detailed instructions on how to build the disk image,
and x86-gapbs-benchmarks-2204.py is a sample script.

After the disk image is built, it can be used as a DiskImageResource in a gem5 simulation.