resources: ARM64 Ubuntu 22.04 Disk Image
This commit adds the files needed to create an ARM64, Ubuntu 22.04 disk image for use as a gem5 resource. The directory qemu_files contains the
scripts for obtaining the disk image and launching it as a QEMU instance. The directory packer_files contains the files for obtaining Packer
and setting up the disk image. README.md contains detailed instructions on how to build the disk image using these files.