resources: Add gfx902 targets to GPU applications
This adds gfx902 as a target for DNNMark's generate_cachefiles.py script
and added the target to the Makefiles.  Note that the Makefiles can be
overridden anyway with the HCC_AMDGPU_TARGET and this is the approach
required to build HACC and PENNANT as the lines in the Makefile are
actually ignored/do not exist. However, by adding the targets, resources
can be built on real hardware with a simple 'make' command.