resources: SPEC 2006 benchmark suite on ARM64 Ubuntu 22.04 Disk Image
This commit adds the files and instructions needed to make a ARM64 Ubuntu 22.04 disk image and install the SPEC 2006 benchmark suite on the disk
image. The qemu-files directory contains the files for getting the disk image and launching it as a QEMU instance. The packer-files directory
contains the files that Packer needs to build the disk image, as well as a script for obtaining Packer. README.md contains detailed instructions
on how to build the disk image and install SPEC 2006 using the scripts and files provided.