resources: GPUFS Ubuntu22.04 ROCm 5.4.2 disk image
Contains everything needed to create a GPUFS disk image from scratch.
Currently we support ROCm 4.2 in the util directory, however everything
up to the currenct ROCm should work. This provides a way to be able to
create a disk image which simulates a more recent GPU software stack.

With this disk image we can run ROCm 5.4 as well as some PyTorch and
Tensorflow applications.

There are instructions to obtain the Linux kernel and it can be
downloaded as a resource.