resources: Add gfx902 targets to GPU applications (#9)
This adds gfx902 as a target for DNNMark's generate_cachefiles.py script
and added the target to the Makefiles. Note that the Makefiles can be
overridden anyway with the HCC_AMDGPU_TARGET and this is the approach
required to build HACC and PENNANT as the lines in the Makefile are
actually ignored/do not exist. However, by adding the targets, resources
can be built on real hardware with a simple 'make' command.

This is related to the discussion on the gem5 main repository about
deprecating GCN3: https://github.com/orgs/gem5/discussions/306 . As a
first step, we want to confirm that a Vega APU target works. That target
is gfx902. I have tested everything in gem5-resources now and everything
works as well as it does on gfx900 (dGPU target). I have uploaded the
scripts I used for testing here:
https://github.com/abmerop/gem5/commit/89f7dba10062eeaf879e489ba76e87dc1baf8df9
. There is no intention of committing these scripts. The commit is
linked solely for reference.