resources: Update GCN3/gfx8 to VEGA/gfx9
The commit does a few things related to simple renames of GCN3. GCN3 is
replaced with Vega in all locations. The targets gfx801 and gfx803 are
replaced with the corresponding gfx9 APU and dGPU targets (gfx902 and
gfx900 respectively). The docker commands update v22-1 with v23-1. This
newer docker image contains an MPI install so that HACC can be built
without a separate docker image. The location of the prebuilt binaries
is updated to a v24-0 path. The old paths contain binaries built with
gfx801, gfx802, and gfx900. They are therefore missing gfx902 and should
be rebuilt before the next release.