resources: Change pannotia fgets to fstream (#24)
There is a bug in either llvm, clang, or hipcc shipped with ROCm 6.0
that causes fgets to hang in gem5, abort in qemu, and seg fault on real
hardware. This seems to be a compiler issue and using fstream instead
solves the problem.

This commit updates all of the graph parser readers to use fstream
instead of fgets/FILE pointers.